DOCAPI/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
azure-storage-blob==12.8.1